+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- proxmox_lxc_template resource.
+
 ### Changed
 - proxmox_lxc.cmds and proxmox_lxc_exec.cmds properties are now sensitive values.